Rename src/render/render_quantum.rs to src/render/quantum.rs
Has no effect in public API, just for our own sanity
